+++ Tests/ValkeyTests/MetricsTests.swift
@@ -26,13 +26,6 @@ struct MetricsTests {
         return factory
     }()
 
-    /// Serializes every metric test across `MetricsTests` and `ClusterMetricsTests`. The two
-    /// suites share `factory` and the process-wide `ValkeyMetrics.pipelineMetrics` /
-    /// `transactionMetrics` handles; `@Suite(.serialized)` only orders tests within a single
-    /// suite, so without this lock the two suites' tests interleave their `factory.reset()`
-    /// and recording calls.
-    static let serializeLock = AsyncTestLock()
-
     private static let primaryAddress = TestStandaloneTopology.Address(host: "127.0.0.1", port: 9100)
 
     @available(valkeySwift 1.0, *)
@@ -67,150 +60,252 @@ struct MetricsTests {
     @Test
     @available(valkeySwift 1.0, *)
     func testSingleCommandSuccessRecordsTimer() async throws {
-        try await Self.serializeLock.withLock {
-            Self.factory.reset()
-            let logger = Logger(label: "test")
-            let topology = await self.makeTopology()
-            let mockConnections = await topology.mock(logger: logger)
-            async let _ = mockConnections.run()
-            try await withClient(mockConnections: mockConnections, logger: logger) { client in
-                try await client.set("foo", value: "Bar")
-                let value = try await client.get("foo")
-                #expect(value.map { String($0) } == "Bar")
-            }
-
-            let label = "valkey.command.get.duration"
-            #expect(Self.factory.timerSamples(label: label, status: "ok").count == 1)
-            #expect(Self.factory.timerSamples(label: label, status: "error").isEmpty)
+        Self.factory.reset()
+        let logger = Logger(label: "test")
+        let topology = await self.makeTopology()
+        let mockConnections = await topology.mock(logger: logger)
+        async let _ = mockConnections.run()
+        try await withClient(mockConnections: mockConnections, logger: logger) { client in
+            try await client.set("foo", value: "Bar")
+            let value = try await client.get("foo")
+            #expect(value.map { String($0) } == "Bar")
         }
+
+        let label = "valkey.command.get.duration"
+        #expect(Self.factory.timerSamples(label: label, status: "ok").count == 1)
+        #expect(Self.factory.timerSamples(label: label, status: "error").isEmpty)
     }
 
     @Test
     @available(valkeySwift 1.0, *)
     func testSingleCommandErrorRecordsErrorStatus() async throws {
-        try await Self.serializeLock.withLock {
-            Self.factory.reset()
-            let logger = Logger(label: "test")
-            let mockConnections = MockServerConnections(logger: logger)
-            await mockConnections.addValkeyServer(.hostname(Self.primaryAddress.host, port: Self.primaryAddress.port)) { command in
-                var iterator = command.makeIterator()
-                switch iterator.next() {
-                case "GET":
-                    return .bulkError("ERR boom")
-                case "ROLE":
-                    return .array([
-                        .bulkString("master"),
-                        .number(1001),
-                        .array([]),
-                    ])
-                default:
-                    return nil
-                }
+        Self.factory.reset()
+        let logger = Logger(label: "test")
+        let mockConnections = MockServerConnections(logger: logger)
+        await mockConnections.addValkeyServer(.hostname(Self.primaryAddress.host, port: Self.primaryAddress.port)) { command in
+            var iterator = command.makeIterator()
+            switch iterator.next() {
+            case "GET":
+                return .bulkError("ERR boom")
+            case "ROLE":
+                return .array([
+                    .bulkString("master"),
+                    .number(1001),
+                    .array([]),
+                ])
+            default:
+                return nil
             }
-            async let _ = mockConnections.run()
-            try await withClient(mockConnections: mockConnections, logger: logger) { client in
-                do {
-                    _ = try await client.get("foo")
-                    Issue.record("expected error")
-                } catch let error as ValkeyClientError {
-                    #expect(error.errorCode == .commandError)
-                }
-            }
-
-            let label = "valkey.command.get.duration"
-            #expect(Self.factory.timerSamples(label: label, status: "error").count == 1)
-            #expect(Self.factory.timerSamples(label: label, status: "ok").isEmpty)
         }
+        async let _ = mockConnections.run()
+        try await withClient(mockConnections: mockConnections, logger: logger) { client in
+            do {
+                _ = try await client.get("foo")
+                Issue.record("expected error")
+            } catch let error as ValkeyClientError {
+                #expect(error.errorCode == .commandError)
+            }
+        }
+
+        let label = "valkey.command.get.duration"
+        #expect(Self.factory.timerSamples(label: label, status: "error").count == 1)
+        #expect(Self.factory.timerSamples(label: label, status: "ok").isEmpty)
     }
 
     @Test
     @available(valkeySwift 1.0, *)
     func testPipelineRecordsTimerAndSize() async throws {
-        try await Self.serializeLock.withLock {
+        Self.factory.reset()
+        let logger = Logger(label: "test")
+        let topology = await self.makeTopology()
+        let mockConnections = await topology.mock(logger: logger)
+        async let _ = mockConnections.run()
+        try await withClient(mockConnections: mockConnections, logger: logger) { client in
+            try await client.set("foo", value: "a")
+            try await client.set("bar", value: "b")
+            // Reset samples so only the pipeline call is measured below.
             Self.factory.reset()
-            let logger = Logger(label: "test")
-            let topology = await self.makeTopology()
-            let mockConnections = await topology.mock(logger: logger)
-            async let _ = mockConnections.run()
-            try await withClient(mockConnections: mockConnections, logger: logger) { client in
-                try await client.set("foo", value: "a")
-                try await client.set("bar", value: "b")
-                // Reset samples so only the pipeline call is measured below.
-                Self.factory.reset()
-                _ = await client.execute(GET("foo"), GET("bar"))
-            }
-
-            #expect(Self.factory.timerSamples(label: "valkey.pipeline.duration", status: nil).count == 1)
-            #expect(Self.factory.recorderSamples(label: "valkey.pipeline.size") == [2.0])
+            _ = await client.execute(GET("foo"), GET("bar"))
         }
+
+        #expect(Self.factory.timerSamples(label: "valkey.pipeline.duration", status: nil).count == 1)
+        #expect(Self.factory.recorderSamples(label: "valkey.pipeline.size") == [2.0])
     }
 
     @Test
     @available(valkeySwift 1.0, *)
     func testTransactionRecordsTimerAndSize() async throws {
-        try await Self.serializeLock.withLock {
-            Self.factory.reset()
-            let logger = Logger(label: "test")
-            let mockConnections = MockServerConnections(logger: logger)
-            // The tests open a single connection so a single shared MULTI/EXEC state is sufficient.
-            let queuedCount = Mutex<Int>(0)
-            let inTransaction = Mutex<Bool>(false)
-            await mockConnections.addValkeyServer(.hostname(Self.primaryAddress.host, port: Self.primaryAddress.port)) { command in
-                var iterator = command.makeIterator()
-                switch iterator.next() {
-                case "MULTI":
-                    inTransaction.withLock { $0 = true }
-                    queuedCount.withLock { $0 = 0 }
-                    return .simpleString("OK")
-                case "EXEC":
-                    let count = queuedCount.withLock { value -> Int in
-                        let count = value
-                        value = 0
-                        return count
-                    }
-                    inTransaction.withLock { $0 = false }
-                    return .array(Array(repeating: .simpleString("OK"), count: count))
-                case "ROLE":
-                    return .array([
-                        .bulkString("master"),
-                        .number(1001),
-                        .array([]),
-                    ])
-                default:
-                    if inTransaction.withLock({ $0 }) {
-                        queuedCount.withLock { $0 += 1 }
-                        return .simpleString("QUEUED")
-                    }
-                    return nil
+        Self.factory.reset()
+        let logger = Logger(label: "test")
+        let mockConnections = MockServerConnections(logger: logger)
+        // The tests open a single connection so a single shared MULTI/EXEC state is sufficient.
+        let queuedCount = Mutex<Int>(0)
+        let inTransaction = Mutex<Bool>(false)
+        await mockConnections.addValkeyServer(.hostname(Self.primaryAddress.host, port: Self.primaryAddress.port)) { command in
+            var iterator = command.makeIterator()
+            switch iterator.next() {
+            case "MULTI":
+                inTransaction.withLock { $0 = true }
+                queuedCount.withLock { $0 = 0 }
+                return .simpleString("OK")
+            case "EXEC":
+                let count = queuedCount.withLock { value -> Int in
+                    let count = value
+                    value = 0
+                    return count
                 }
+                inTransaction.withLock { $0 = false }
+                return .array(Array(repeating: .simpleString("OK"), count: count))
+            case "ROLE":
+                return .array([
+                    .bulkString("master"),
+                    .number(1001),
+                    .array([]),
+                ])
+            default:
+                if inTransaction.withLock({ $0 }) {
+                    queuedCount.withLock { $0 += 1 }
+                    return .simpleString("QUEUED")
+                }
+                return nil
             }
-            async let _ = mockConnections.run()
-            try await withClient(mockConnections: mockConnections, logger: logger) { client in
-                _ = try await client.transaction(SET("foo", value: "10"), INCR("foo"))
-            }
-
-            #expect(Self.factory.timerSamples(label: "valkey.transaction.duration", status: nil).count == 1)
-            #expect(Self.factory.recorderSamples(label: "valkey.transaction.size") == [2.0])
         }
+        async let _ = mockConnections.run()
+        try await withClient(mockConnections: mockConnections, logger: logger) { client in
+            _ = try await client.transaction(SET("foo", value: "10"), INCR("foo"))
+        }
+
+        #expect(Self.factory.timerSamples(label: "valkey.transaction.duration", status: nil).count == 1)
+        #expect(Self.factory.recorderSamples(label: "valkey.transaction.size") == [2.0])
     }
 
     @Test
     @available(valkeySwift 1.0, *)
     func testMetricsDisabledSkipsEmission() async throws {
-        try await Self.serializeLock.withLock {
-            Self.factory.reset()
-            let logger = Logger(label: "test")
-            let topology = await self.makeTopology()
-            let mockConnections = await topology.mock(logger: logger)
-            async let _ = mockConnections.run()
-            try await withClient(mockConnections: mockConnections, metricsEnabled: false, logger: logger) { client in
-                try await client.set("foo", value: "Bar")
-                _ = try await client.get("foo")
-            }
+        Self.factory.reset()
+        let logger = Logger(label: "test")
+        let topology = await self.makeTopology()
+        let mockConnections = await topology.mock(logger: logger)
+        async let _ = mockConnections.run()
+        try await withClient(mockConnections: mockConnections, metricsEnabled: false, logger: logger) { client in
+            try await client.set("foo", value: "Bar")
+            _ = try await client.get("foo")
+        }
 
-            let label = "valkey.command.get.duration"
-            #expect(Self.factory.timerSamples(label: label, status: "ok").isEmpty)
-            #expect(Self.factory.timerSamples(label: label, status: "error").isEmpty)
+        let label = "valkey.command.get.duration"
+        #expect(Self.factory.timerSamples(label: label, status: "ok").isEmpty)
+        #expect(Self.factory.timerSamples(label: label, status: "error").isEmpty)
+    }
+
+    /// Nested suite for cluster-client metrics. Lives inside `MetricsTests` so it inherits the
+    /// outer `.serialized` trait — that single ordering is what keeps cluster and standalone
+    /// tests from interleaving on the shared `factory` and the process-wide
+    /// `ValkeyMetrics.pipelineMetrics` / `transactionMetrics` handles.
+    @Suite
+    struct Cluster {
+        private var sixNodeHealthyCluster: TestCluster {
+            get async {
+                await TestCluster(shards: [
+                    TestCluster.Shard(
+                        hashKeyRanges: [0...5460],
+                        primary: .init(host: "127.0.0.1", port: 17000),
+                        replicas: [.init(host: "127.0.0.1", port: 17001)]
+                    ),
+                    TestCluster.Shard(
+                        hashKeyRanges: [5461...10922],
+                        primary: .init(host: "127.0.0.1", port: 17002),
+                        replicas: [.init(host: "127.0.0.1", port: 17003)]
+                    ),
+                    TestCluster.Shard(
+                        hashKeyRanges: [10923...16383],
+                        primary: .init(host: "127.0.0.1", port: 17004),
+                        replicas: [.init(host: "127.0.0.1", port: 17005)]
+                    ),
+                ])
+            }
+        }
+
+        @available(valkeySwift 1.0, *)
+        private func withClient(
+            mockConnections: MockServerConnections,
+            logger: Logger,
+            operation: @escaping @Sendable (ValkeyClusterClient) async throws -> Void
+        ) async throws {
+            var clientConfig = ValkeyClientConfiguration(readOnlyCommandNodeSelection: .cycleReplicas)
+            clientConfig.metrics.enabled = true
+            let client = ValkeyClusterClient(
+                nodeDiscovery: ValkeyStaticNodeDiscovery([.init(endpoint: "127.0.0.1", port: 17000)]),
+                configuration: .init(client: clientConfig),
+                eventLoopGroup: mockConnections.eventLoop,
+                logger: logger,
+                channelFactory: mockConnections.connectionManagerCustomHandler
+            )
+            try await withThrowingTaskGroup(of: Void.self) { group in
+                group.addTask { await client.run() }
+                group.addTask { try await operation(client) }
+                try await group.next()
+                group.cancelAll()
+            }
+        }
+
+        /// Pipeline that hits a MOVED redirect (slot migrated to another shard) must produce exactly
+        /// one pipeline metric sample per user-level call regardless of the per-node retry happening
+        /// inside the cluster client.
+        @Test
+        @available(valkeySwift 1.0, *)
+        func testClusterPipelineWithMovedRetryRecordsOnce() async throws {
+            MetricsTests.factory.reset()
+            var logger = Logger(label: "Valkey")
+            logger.logLevel = .debug
+            let cluster = await self.sixNodeHealthyCluster
+            let mockConnections = await cluster.mock(logger: logger)
+            async let _ = mockConnections.run()
+            try await withClient(mockConnections: mockConnections, logger: logger) { client in
+                try await client.set("randomKey", value: "before")
+                // Migrate the slot for "randomKey" to shard 2 so the next pipeline gets MOVED
+                // on shard 0 and the cluster client retries against the new owner.
+                let hashSlot = HashSlot(key: "randomKey".utf8).rawValue
+                await cluster.migrateSlots(hashSlot...hashSlot, to: 2)
+
+                MetricsTests.factory.reset()
+                let results = await client.execute(
+                    GET("randomKey"),
+                    SET("randomKey", value: "after"),
+                    GET("randomKey")
+                )
+                try #expect(results.0.get().map { String($0) } == "before")
+                try #expect(results.2.get().map { String($0) } == "after")
+
+                #expect(MetricsTests.factory.timerSamples(label: "valkey.pipeline.duration", status: nil).count == 1)
+                #expect(MetricsTests.factory.recorderSamples(label: "valkey.pipeline.size") == [3.0])
+            }
+        }
+
+        /// Same invariant for transactions: a single user-level transaction call produces exactly one
+        /// transaction metric sample regardless of MOVED retries.
+        @Test
+        @available(valkeySwift 1.0, *)
+        func testClusterTransactionWithMovedRetryRecordsOnce() async throws {
+            MetricsTests.factory.reset()
+            var logger = Logger(label: "Valkey")
+            logger.logLevel = .debug
+            let cluster = await self.sixNodeHealthyCluster
+            let mockConnections = await cluster.mock(logger: logger)
+            async let _ = mockConnections.run()
+            try await withClient(mockConnections: mockConnections, logger: logger) { client in
+                try await client.set("txnKey", value: "before")
+                let hashSlot = HashSlot(key: "txnKey".utf8).rawValue
+                await cluster.migrateSlots(hashSlot...hashSlot, to: 2)
+
+                MetricsTests.factory.reset()
+                _ = try await client.transaction(
+                    SET("txnKey", value: "v1"),
+                    SET("txnKey", value: "v2")
+                )
+
+                #expect(MetricsTests.factory.timerSamples(label: "valkey.transaction.duration", status: nil).count == 1)
+                #expect(MetricsTests.factory.recorderSamples(label: "valkey.transaction.size") == [2.0])
+            }
         }
     }
 
@@ -372,42 +467,5 @@ final class NoOpMeter: MeterHandler, Sendable {
     func set(_ value: Double) {}
     func increment(by: Double) {}
     func decrement(by: Double) {}
-}
-
-/// Single-permit async semaphore used to serialize entire async test bodies across suites.
-actor AsyncTestLock {
-    private var available = true
-    private var waiters: [CheckedContinuation<Void, Never>] = []
-
-    func wait() async {
-        if self.available {
-            self.available = false
-            return
-        }
-        await withCheckedContinuation { cont in
-            self.waiters.append(cont)
-        }
-    }
-
-    func signal() {
-        if let next = self.waiters.first {
-            self.waiters.removeFirst()
-            next.resume()
-        } else {
-            self.available = true
-        }
-    }
-
-    func withLock<T: Sendable>(_ body: @Sendable () async throws -> T) async rethrows -> T {
-        await self.wait()
-        do {
-            let result = try await body()
-            self.signal()
-            return result
-        } catch {
-            self.signal()
-            throw error
-        }
-    }
 }
 #endif
